+++ src/Symfony/Component/Form/FormConfig.php
@@ -121,9 +121,10 @@ class FormConfig implements FormConfigEditorInterface
     /**
      * Creates an empty form configuration.
      *
-     * @param  string                   $name       The form name.
-     * @param  string                   $dataClass  The class of the form's data.
-     * @param  EventDispatcherInterface $dispatcher The event dispatcher.
+     * @param string                   $name       The form name
+     * @param string                   $dataClass  The class of the form's data
+     * @param EventDispatcherInterface $dispatcher The event dispatcher
+     * @param array                    $options    The form options
      *
      * @throws UnexpectedTypeException   If the name is not a string.
      * @throws \InvalidArgumentException If the data class is not a valid class or if
@@ -178,9 +179,13 @@ class FormConfig implements FormConfigEditorInterface
     /**
      * {@inheritdoc}
      */
-    public function addViewTransformer(DataTransformerInterface $viewTransformer)
+    public function addViewTransformer(DataTransformerInterface $viewTransformer, $forcePrepend = false)
     {
-        $this->viewTransformers[] = $viewTransformer;
+        if ($forcePrepend) {
+            array_unshift($this->viewTransformers, $viewTransformer);
+        } else {
+            $this->viewTransformers[] = $viewTransformer;
+        }
 
         return $this;
     }
@@ -221,9 +226,7 @@ class FormConfig implements FormConfigEditorInterface
      */
     public function prependClientTransformer(DataTransformerInterface $viewTransformer)
     {
-        array_unshift($this->viewTransformers, $viewTransformer);
-
-        return $this;
+        return $this->addViewTransformer($viewTransformer, true);
     }
 
     /**
@@ -242,9 +245,13 @@ class FormConfig implements FormConfigEditorInterface
     /**
      * {@inheritdoc}
      */
-    public function addModelTransformer(DataTransformerInterface $modelTransformer)
+    public function addModelTransformer(DataTransformerInterface $modelTransformer, $forceAppend = false)
     {
-        array_unshift($this->modelTransformers, $modelTransformer);
+        if ($forceAppend) {
+            $this->modelTransformers[] = $modelTransformer;
+        } else {
+            array_unshift($this->modelTransformers, $modelTransformer);
+        }
 
         return $this;
     }
@@ -270,9 +277,7 @@ class FormConfig implements FormConfigEditorInterface
      */
     public function appendNormTransformer(DataTransformerInterface $modelTransformer)
     {
-        $this->modelTransformers[] = $modelTransformer;
-
-        return $this;
+        return $this->addModelTransformer($modelTransformer, true);
     }
 
     /**