+++ src/Symfony/Component/Form/Tests/Extension/Core/Type/DateTypeTest.php
@@ -12,6 +12,7 @@
 namespace Symfony\Component\Form\Tests\Extension\Core\Type;
 
 use Symfony\Component\Form\Extension\Core\View\ChoiceView;
+use Symfony\Component\Form\FormError;
 
 class DateTypeTest extends LocalizedTestCase
 {